Extract shared functionality into a "core" app
It's useful to structure this as a django app because I want to put some
model stuff in there, and models shouldn't be outside of a Django app.

In a follow up commit, I will be extending the base model to include
auditing.
