+++ manage_breast_screening/clinics/models.py
@@ -1,17 +1,9 @@
-import uuid
 from datetime import date
 from enum import StrEnum
 
 from django.db import models
 
-
-class BaseModel(models.Model):
-    class Meta:
-        abstract = True
-
-    id = models.UUIDField(default=uuid.uuid4, editable=False, primary_key=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
+from ..core.models import BaseModel
 
 
 class Provider(BaseModel):
