+++ manage_breast_screening/core/jinja2/layout-app.jinja
